refactor(node): split metrics responsibilities
