+++ crates/node/src/metrics.rs
@@ -1,934 +1,23 @@
 extern crate alloc;
 
-use alloc::collections::BTreeMap;
-use alloc::string::String;
-use alloc::sync::Arc;
-use alloc::vec::Vec;
-use metrics_exporter_prometheus::{PrometheusBuilder, PrometheusHandle};
-use std::io::{Read, Write};
-use std::net::{SocketAddr, TcpListener, TcpStream};
-use std::sync::OnceLock;
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::thread::{self, JoinHandle};
-use std::time::{Duration, Instant};
+mod evidence;
+mod prometheus;
+mod uptime;
+mod warnings;
 
-use anyhow::Result;
-use parking_lot::Mutex;
+pub use evidence::{
+    Cell, CorpusIdentity, EvidenceError, EvidenceIdentity, Interval, IntervalKind, LEDGER_SCHEMA,
+    Ledger, Sample, Sha256Hex,
+};
+pub use prometheus::MetricsServer;
+pub(crate) use prometheus::start_metrics;
+pub use uptime::{process_start, process_uptime, record_process_start};
+pub use warnings::{WarningKind, Warnings, node_warnings};
 
-/// Monotonic process start, recorded once at node startup.
-///
-/// Core initializes its uptime clock at process start (`GetUptime`,
-/// `src/common/system.cpp`), so every later query reports true elapsed
-/// runtime. A lazily initialized clock would restart on first read, which is
-/// the first-call-reports-approximately-zero bug the utility parity audit
-/// records for `uptime`; the earliest-wins record here makes the node, not
-/// the first RPC caller, own the origin.
-static PROCESS_START: OnceLock<Instant> = OnceLock::new();
-
-/// Records the authoritative process start instant for uptime accounting.
-///
-/// Idempotent and earliest-wins: the first caller fixes the uptime origin,
-/// mirroring Core's static initialization. `run` records this before wiring
-/// subsystems so the clock covers the whole node lifecycle.
-pub fn record_process_start(start: Instant) {
-    let _ = PROCESS_START.set(start);
-}
-
-/// Returns the recorded process start instant, or `None` before
-/// [`record_process_start`] has run.
-#[must_use]
-pub fn process_start() -> Option<Instant> {
-    PROCESS_START.get().copied()
-}
-
-/// Returns monotonic uptime since the recorded process start, or `None`
-/// before [`record_process_start`] has run.
-///
-/// `None` keeps the not-yet-started state observable instead of silently
-/// restarting the clock on read.
-#[must_use]
-pub fn process_uptime() -> Option<Duration> {
-    PROCESS_START.get().map(Instant::elapsed)
-}
-
-/// Warning kinds the node can raise, mirroring Core's kernel and node warning
-/// ids.
-///
-/// Declaration order is the report order: Core keys its registry by
-/// `std::variant<kernel::Warning, node::Warning>`, so kernel-issued warnings
-/// sort before node-issued ones.
-#[derive(Clone, Copy, Debug, Eq, Ord, PartialEq, PartialOrd)]
-pub enum WarningKind {
-    /// A versionbit unknown to this node activated on the network.
-    UnknownNewRulesActivated,
-    /// An invalid chain with substantially more work than the best chain was
-    /// found.
-    LargeWorkInvalidChain,
-    /// The system clock is far out of sync with peer clocks.
-    ClockOutOfSync,
-    /// An internal error put the node in a state it cannot recover from.
-    FatalInternalError,
-}
-
-/// Node warning registry, mirroring Core's `node::Warnings`.
-///
-/// At most one message per [`WarningKind`]: setting an already-active kind
-/// keeps the original message, and [`Warnings::messages`] reports in kind
-/// order. The registry is node-owned state; RPC projections read it live at
-/// call time instead of copying warnings into transport-owned storage.
-pub struct Warnings {
-    active: Mutex<BTreeMap<WarningKind, String>>,
-}
-
-impl Default for Warnings {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl Warnings {
-    /// Creates an empty registry.
-    pub const fn new() -> Self {
-        Self {
-            active: Mutex::new(BTreeMap::new()),
-        }
-    }
-
-    /// Activates the warning for `kind` with `message`.
-    ///
-    /// Returns `true` when the warning was newly set. An already-active kind
-    /// keeps its original message and returns `false`, matching Core's
-    /// `Warnings::Set`.
-    pub fn set(&self, kind: WarningKind, message: impl Into<String>) -> bool {
-        let mut active = self.active.lock();
-        if active.contains_key(&kind) {
-            return false;
-        }
-        active.insert(kind, message.into());
-        true
-    }
-
-    /// Deactivates the warning for `kind`, returning whether one was active.
-    pub fn unset(&self, kind: WarningKind) -> bool {
-        self.active.lock().remove(&kind).is_some()
-    }
-
-    /// Returns the active warning messages ordered by kind.
-    #[must_use]
-    pub fn messages(&self) -> Vec<String> {
-        self.active.lock().values().cloned().collect()
-    }
-}
-
-static NODE_WARNINGS: Warnings = Warnings::new();
-
-/// Returns the process-wide node warning registry.
-///
-/// This is the authoritative warnings fact source: the `getblockchaininfo`,
-/// `getnetworkinfo`, and `getmininginfo` projections read it at invocation,
-/// the way Core reads `NodeContext::warnings` through `GetWarningsForRpc`.
-#[must_use]
-pub fn node_warnings() -> &'static Warnings {
-    &NODE_WARNINGS
-}
-
-fn describe_node_metrics() {
-    metrics::describe_counter!("node.event_loop.mempool_ticks", "mempool maintenance ticks");
-    metrics::describe_counter!("node.event_loop.metrics_scrapes", "metrics scrape ticks");
-    metrics::describe_counter!("node.event_loop.sync_ticks", "block sync ticks");
-    metrics::describe_counter!(
-        "node.event_loop.sync_wakes",
-        "block sync wakeups from inbound p2p data"
-    );
-    metrics::describe_gauge!(
-        "node.shutdown.requested",
-        "whether shutdown has been requested"
-    );
-    metrics::describe_histogram!(
-        "node.event_loop.tick_seconds",
-        "event loop tick latency seconds"
-    );
-    metrics::describe_counter!(
-        "node.sync.duplicate_deliveries",
-        "blocks received that were already staged"
-    );
-    metrics::describe_histogram!(
-        "node.sync.apply_idle_seconds",
-        "durations the apply frontier stayed starved while the window owed downloads"
-    );
-    metrics::describe_histogram!(
-        "node.sync.download_blocked_by_apply_seconds",
-        "durations the window front stayed in flight while apply held the frontier"
-    );
-    metrics::describe_gauge!(
-        "node.sync.pending_blocks_high_water",
-        "highest in-flight block count observed"
-    );
-    metrics::describe_gauge!(
-        "node.sync.pending_bytes_high_water",
-        "highest in-flight byte estimate observed"
-    );
-    metrics::describe_gauge!(
-        "node.sync.staged_blocks_high_water",
-        "highest staged block count observed"
-    );
-    metrics::describe_gauge!(
-        "node.sync.staged_bytes_high_water",
-        "highest staged byte total observed"
-    );
-    metrics::describe_counter!(
-        "storage.writes_total",
-        "storage write batches applied, by backend and durability"
-    );
-    metrics::describe_counter!(
-        "storage.flushes_total",
-        "storage durability flushes by backend"
-    );
-    metrics::describe_histogram!("storage.write_bytes", "storage write batch payload bytes");
-    metrics::describe_gauge!(
-        "storage.cache_capacity_bytes",
-        "configured per-engine cache capacity in bytes"
-    );
-}
-
-static PROMETHEUS_HANDLE: Mutex<Option<(EvidenceIdentity, PrometheusHandle)>> = Mutex::new(None);
-
-/// One process serves one identity: every scraped sample carries the
-/// artifact, configuration, corpus and durability it was taken under as
-/// global labels, so a reader can never attribute a value to the wrong build.
-fn prometheus_handle(identity: &EvidenceIdentity) -> Result<PrometheusHandle> {
-    let mut slot = PROMETHEUS_HANDLE.lock();
-    if let Some((installed, handle)) = slot.as_ref() {
-        anyhow::ensure!(
-            installed == identity,
-            "metrics recorder already serves a different evidence identity"
-        );
-        return Ok(handle.clone());
-    }
-    let mut builder = PrometheusBuilder::new();
-    for (label, value) in identity.labels() {
-        builder = builder.add_global_label(label, value);
-    }
-    let handle = builder
-        .install_recorder()
-        .map_err(|error| anyhow::anyhow!("install prometheus recorder: {error}"))?;
-    *slot = Some((identity.clone(), handle.clone()));
-    Ok(handle)
-}
-
-/// Process-global Prometheus scrape listener bound by [`start_metrics`].
-pub struct MetricsServer {
-    local_addr: SocketAddr,
-    stop: Arc<AtomicBool>,
-    thread: Option<JoinHandle<()>>,
-}
-
-impl MetricsServer {
-    /// Binds `addr` before installing the recorder, then serves Prometheus text.
-    ///
-    /// Listener-first ordering keeps an occupied-address failure from consuming
-    /// the process-global recorder slot, so a later in-process retry cannot hit
-    /// `SetRecorderError`.
-    pub fn bind(
-        addr: SocketAddr,
-        shutdown: Arc<AtomicBool>,
-        identity: &EvidenceIdentity,
-    ) -> Result<Self> {
-        let listener = TcpListener::bind(addr)?;
-        let local_addr = listener.local_addr()?;
-        let handle = prometheus_handle(identity)?;
-        describe_node_metrics();
-        listener.set_nonblocking(true)?;
-        let stop = Arc::new(AtomicBool::new(false));
-        let thread_stop = Arc::clone(&stop);
-        let thread = thread::Builder::new()
-            .name("bitcoin-rs-metrics".into())
-            .spawn(move || serve_metrics(&listener, &handle, &thread_stop, &shutdown))?;
-        Ok(Self {
-            local_addr,
-            stop,
-            thread: Some(thread),
-        })
-    }
-
-    /// Address the scrape thread is listening on.
-    #[must_use]
-    pub const fn local_addr(&self) -> SocketAddr {
-        self.local_addr
-    }
-
-    /// Signals the scrape thread and waits for it to exit.
-    pub fn join(mut self) {
-        self.stop_and_join();
-    }
-
-    fn stop_and_join(&mut self) {
-        self.stop.store(true, Ordering::Release);
-        if let Some(thread) = self.thread.take() {
-            let _ = thread.join();
-        }
-    }
-}
-
-impl Drop for MetricsServer {
-    fn drop(&mut self) {
-        self.stop_and_join();
-    }
-}
-
-/// Starts the production scrape listener when `metrics_bind` is configured.
-///
-/// This is the entry `run` uses after [`crate::state::NodeState::open`].
-pub(crate) fn start_metrics(
-    bind: Option<SocketAddr>,
-    shutdown: Arc<AtomicBool>,
-    identity: &EvidenceIdentity,
-) -> Result<Option<MetricsServer>> {
-    bind.map(|addr| MetricsServer::bind(addr, shutdown, identity))
-        .transpose()
-}
-
-fn serve_metrics(
-    listener: &TcpListener,
-    handle: &PrometheusHandle,
-    stop: &Arc<AtomicBool>,
-    shutdown: &Arc<AtomicBool>,
-) {
-    loop {
-        if stop.load(Ordering::Acquire) || shutdown.load(Ordering::Acquire) {
-            break;
-        }
-        match listener.accept() {
-            Ok((mut stream, _)) => {
-                let _ = stream.set_nonblocking(false);
-                let _ = stream.set_read_timeout(Some(Duration::from_secs(2)));
-                let _ = stream.set_write_timeout(Some(Duration::from_secs(2)));
-                serve_scrape(&mut stream, handle);
-            }
-            Err(error) if error.kind() == std::io::ErrorKind::WouldBlock => {
-                thread::sleep(Duration::from_millis(20));
-            }
-            Err(error) if error.kind() == std::io::ErrorKind::Interrupted => {}
-            Err(_) => break,
-        }
-    }
-}
-
-fn serve_scrape(stream: &mut TcpStream, handle: &PrometheusHandle) {
-    let mut buf = [0_u8; 1024];
-    let _ = stream.read(&mut buf);
-    let body = handle.render();
-    let response = format!(
-        "HTTP/1.1 200 OK\r\nContent-Type: text/plain; version=0.0.4; charset=utf-8\r\nContent-Length: {}\r\nConnection: close\r\n\r\n{body}",
-        body.len(),
-    );
-    let _ = stream.write_all(response.as_bytes());
-    let _ = stream.flush();
-}
-
-/// A SHA-256 digest carried as 64 lowercase hex characters in evidence.
-///
-/// A digest is bytes, not a label: a placeholder such as "unmeasured" cannot
-/// parse, so an identity is either real or absent.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct Sha256Hex(pub [u8; 32]);
-
-impl Sha256Hex {
-    /// Hashes `bytes` with SHA-256.
-    #[must_use]
-    pub fn digest(bytes: &[u8]) -> Self {
-        use sha2::Digest as _;
-        Self(sha2::Sha256::digest(bytes).into())
-    }
-}
-
-impl core::fmt::Display for Sha256Hex {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        for byte in self.0 {
-            write!(f, "{byte:02x}")?;
-        }
-        Ok(())
-    }
-}
-
-impl core::str::FromStr for Sha256Hex {
-    type Err = EvidenceError;
-
-    fn from_str(text: &str) -> Result<Self, EvidenceError> {
-        let malformed = || EvidenceError::MalformedDigest(text.into());
-        if text.len() != 64 {
-            return Err(malformed());
-        }
-        let mut bytes = [0_u8; 32];
-        for (byte, pair) in bytes.iter_mut().zip(text.as_bytes().as_chunks::<2>().0) {
-            let text = core::str::from_utf8(pair).map_err(|_| malformed())?;
-            if text.bytes().any(|c| c.is_ascii_uppercase()) {
-                return Err(malformed());
-            }
-            *byte = u8::from_str_radix(text, 16).map_err(|_| malformed())?;
-        }
-        Ok(Self(bytes))
-    }
-}
-
-impl serde::Serialize for Sha256Hex {
-    fn serialize<S: serde::Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        serializer.collect_str(self)
-    }
-}
-
-impl<'de> serde::Deserialize<'de> for Sha256Hex {
-    fn deserialize<D: serde::Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        let text = String::deserialize(deserializer)?;
-        text.parse().map_err(serde::de::Error::custom)
-    }
-}
-
-/// The corpus a measurement replayed.
-#[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct CorpusIdentity {
-    /// Corpus identifier from `docs/contracts/campaign-corpora.md`.
-    pub id: String,
-    /// Digest of the corpus manifest.
-    pub manifest_sha256: Sha256Hex,
-}
-
-/// Everything a measurement was taken under.
-///
-/// A number without this record is a rumor: it cannot be matched against a
-/// control cell or regenerated later.
-#[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct EvidenceIdentity {
-    /// Digest of the executable that produced the sample.
-    pub binary_sha256: Sha256Hex,
-    /// Crate version of that executable.
-    pub version: String,
-    /// Digest of the fully resolved configuration.
-    pub config_sha256: Sha256Hex,
-    /// Replayed corpus. A live node has none; a product cell always has one.
-    pub corpus: Option<CorpusIdentity>,
-    /// Storage backend that held the state.
-    pub backend: String,
-    /// Durability policy in force, for example `journal:500b/5s`.
-    pub durability: String,
-    /// Hardware the sample ran on: CPU model and logical core count.
-    pub hardware: String,
-}
-
-/// CPU model and core count, the two hardware facts a matched treatment
-/// must share before its numbers are comparable.
-fn hardware_identity() -> String {
-    let model = std::fs::read_to_string("/proc/cpuinfo")
-        .ok()
-        .and_then(|info| {
-            info.lines()
-                .find_map(|line| line.strip_prefix("model name"))
-                .and_then(|rest| rest.split_once(':'))
-                .map(|(_, model)| model.trim().to_owned())
-        })
-        .unwrap_or_else(|| "unknown-cpu".into());
-    let cores = std::thread::available_parallelism().map_or(0, usize::from);
-    format!("{model} x{cores}")
-}
-
-impl EvidenceIdentity {
-    /// Identity for the running process under `config`.
-    ///
-    /// The digest covers the executable file on disk and the resolved
-    /// configuration's debug rendering, which is the one canonical form the
-    /// runtime already owns.
-    pub fn of_process(config: &crate::config::NodeConfig) -> Result<Self> {
-        let executable = std::env::current_exe()?;
-        let binary_sha256 = Sha256Hex::digest(&std::fs::read(&executable)?);
-        let journal = config.chainstate_journal;
-        let durability = if journal.enabled {
-            format!("journal:{}b/{}s", journal.blocks, journal.seconds)
-        } else {
-            "checkpoint-only".into()
-        };
-        Ok(Self {
-            binary_sha256,
-            version: env!("CARGO_PKG_VERSION").into(),
-            config_sha256: Sha256Hex::digest(format!("{config:?}").as_bytes()),
-            corpus: None,
-            backend: config.storage.backend.as_str().into(),
-            durability,
-            hardware: hardware_identity(),
-        })
-    }
-
-    /// The identity as Prometheus global labels.
-    #[must_use]
-    pub fn labels(&self) -> Vec<(&'static str, String)> {
-        let mut labels = vec![
-            ("binary_sha256", self.binary_sha256.to_string()),
-            ("version", self.version.clone()),
-            ("config_sha256", self.config_sha256.to_string()),
-            ("backend", self.backend.clone()),
-            ("durability", self.durability.clone()),
-            ("hardware", self.hardware.clone()),
-        ];
-        if let Some(corpus) = &self.corpus {
-            labels.push(("corpus_id", corpus.id.clone()));
-            labels.push(("corpus_manifest_sha256", corpus.manifest_sha256.to_string()));
-        }
-        labels
-    }
-}
-
-/// Where a timed interval sits relative to the measured product.
-#[derive(Clone, Copy, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum IntervalKind {
-    /// Inside the node process.
-    Inside,
-    /// Outside the node process, for example harness setup.
-    Outside,
-    /// A duration the domain defines without wall-clock placement.
-    DomainDefined,
-}
-
-/// A half-open timed interval in nanoseconds on the run's clock.
-#[derive(Clone, Copy, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct Interval {
-    /// Placement of the interval.
-    pub kind: IntervalKind,
-    /// Start offset.
-    pub start_ns: u64,
-    /// End offset, never before the start.
-    pub end_ns: u64,
-}
-
-impl Interval {
-    fn check(self) -> Result<(), EvidenceError> {
-        if self.end_ns < self.start_ns {
-            return Err(EvidenceError::InvertedInterval(self));
-        }
-        Ok(())
-    }
-
-    /// True when the intervals share any instant, which covers both nesting
-    /// and concurrency.
-    #[must_use]
-    pub fn overlaps(self, other: Self) -> bool {
-        self.start_ns < other.end_ns && other.start_ns < self.end_ns
-    }
-}
-
-/// One measurement of one cell by one owner.
-#[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct Sample {
-    /// Ledger path the interval measures, for example `cell.wall`.
-    pub path: String,
-    /// Component that owns the measured resources, for example `node`.
-    pub owner: String,
-    /// Identity the sample was taken under.
-    pub identity: EvidenceIdentity,
-    /// Timed interval the resources were consumed in.
-    pub interval: Interval,
-    /// CPU time consumed, when the owner measured it.
-    pub cpu_ns: Option<u64>,
-    /// Wall time consumed.
-    pub elapsed_ns: u64,
-    /// Peak resident set, when the owner measured it.
-    pub rss_peak_bytes: Option<u64>,
-    /// Bytes read plus written, when the owner measured it.
-    pub io_bytes: Option<u64>,
-    /// Physical storage held at the end of the interval, when measured.
-    pub storage_bytes: Option<u64>,
-}
-
-/// Combines two optional measurements; an absent side cannot silently absorb
-/// a present one, because that would collapse coverage into a number.
-fn combine(
-    left: Option<u64>,
-    right: Option<u64>,
-    join: impl Fn(u64, u64) -> Option<u64>,
-) -> Result<Option<u64>, EvidenceError> {
-    match (left, right) {
-        (None, None) => Ok(None),
-        (Some(left), Some(right)) => join(left, right).map(Some).ok_or(EvidenceError::Overflow),
-        _ => Err(EvidenceError::MismatchedTreatment),
-    }
-}
-
-impl Sample {
-    fn check(&self) -> Result<(), EvidenceError> {
-        if self.identity.corpus.is_none() {
-            return Err(EvidenceError::MissingCorpus);
-        }
-        self.interval.check()
-    }
-
-    /// Adds two disjoint samples of the same owner and identity.
-    ///
-    /// Nested and concurrent intervals share instants, so their resources
-    /// were consumed once and cannot be added; extrema take the maximum.
-    pub fn sum(&self, other: &Self) -> Result<Self, EvidenceError> {
-        if self.path != other.path || self.owner != other.owner || self.identity != other.identity {
-            return Err(EvidenceError::MismatchedTreatment);
-        }
-        if self.interval.overlaps(other.interval) {
-            return Err(EvidenceError::OverlappingIntervals(
-                self.interval,
-                other.interval,
-            ));
-        }
-        Ok(Self {
-            path: self.path.clone(),
-            owner: self.owner.clone(),
-            identity: self.identity.clone(),
-            interval: Interval {
-                kind: self.interval.kind,
-                start_ns: self.interval.start_ns.min(other.interval.start_ns),
-                end_ns: self.interval.end_ns.max(other.interval.end_ns),
-            },
-            cpu_ns: combine(self.cpu_ns, other.cpu_ns, u64::checked_add)?,
-            elapsed_ns: self
-                .elapsed_ns
-                .checked_add(other.elapsed_ns)
-                .ok_or(EvidenceError::Overflow)?,
-            rss_peak_bytes: combine(self.rss_peak_bytes, other.rss_peak_bytes, |a, b| {
-                Some(a.max(b))
-            })?,
-            io_bytes: combine(self.io_bytes, other.io_bytes, u64::checked_add)?,
-            storage_bytes: combine(self.storage_bytes, other.storage_bytes, |a, b| {
-                Some(a.max(b))
-            })?,
-        })
-    }
-}
-
-/// One product cell and every sample ever recorded for it.
-///
-/// An empty history is the honest state of an unmeasured cell; it is kept,
-/// never dropped or defaulted.
-#[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct Cell {
-    /// Cell identifier from the ledger matrix.
-    pub id: String,
-    /// Append-only sample history.
-    pub samples: Vec<Sample>,
-}
-
-/// Schema tag every ledger and evidence file must carry.
-pub const LEDGER_SCHEMA: &str = "bitcoin-rs-hot-path-ledger-v2";
-
-/// The cell histories of `docs/benchmarks/hot-path-ledger.toml`.
-///
-/// Other top-level tables of that file belong to the attribution contract and
-/// pass through untouched; this type owns only what a run may append to.
-#[derive(Clone, Debug, PartialEq, serde::Serialize, serde::Deserialize)]
-pub struct Ledger {
-    /// Must equal [`LEDGER_SCHEMA`].
-    pub schema: String,
-    /// Every declared cell, measured or not.
-    pub cells: Vec<Cell>,
-    /// Top-level attribution-contract tables owned outside the benchmark tool.
-    #[serde(flatten)]
-    pub contract: std::collections::BTreeMap<String, toml::Value>,
-}
-
-impl Ledger {
-    /// Parses ledger TOML and rejects any sample that lacks a full identity.
-    pub fn parse(text: &str) -> Result<Self, EvidenceError> {
-        let ledger: Self =
-            toml::from_str(text).map_err(|error| EvidenceError::Schema(error.to_string()))?;
-        if ledger.schema != LEDGER_SCHEMA {
-            return Err(EvidenceError::Schema(format!(
-                "schema {} is not {LEDGER_SCHEMA}",
-                ledger.schema
-            )));
-        }
-        for cell in &ledger.cells {
-            for sample in &cell.samples {
-                sample.check()?;
-            }
-        }
-        Ok(ledger)
-    }
-
-    /// Appends a sample to a declared cell. Repeats are kept: a second run is
-    /// evidence, not a correction.
-    pub fn record(&mut self, cell: &str, sample: Sample) -> Result<(), EvidenceError> {
-        sample.check()?;
-        let cell = self
-            .cells
-            .iter_mut()
-            .find(|candidate| candidate.id == cell)
-            .ok_or_else(|| EvidenceError::UnknownCell(cell.into()))?;
-        cell.samples.push(sample);
-        Ok(())
-    }
-
-    /// Renders the ledger as TOML.
-    pub fn render(&self) -> Result<String, EvidenceError> {
-        toml::to_string(self).map_err(|error| EvidenceError::Schema(error.to_string()))
-    }
-}
-
-/// Why evidence was refused.
-#[derive(Debug, thiserror::Error, Eq, PartialEq)]
-pub enum EvidenceError {
-    /// A digest was not 64 lowercase hex characters.
-    #[error("digest {0:?} is not 64 lowercase hex characters")]
-    MalformedDigest(String),
-    /// The ledger text did not match the schema.
-    #[error("ledger schema: {0}")]
-    Schema(String),
-    /// A product sample named no corpus.
-    #[error("product sample carries no corpus identity")]
-    MissingCorpus,
-    /// An interval ended before it started.
-    #[error("interval {0:?} ends before it starts")]
-    InvertedInterval(Interval),
-    /// The two samples were not taken under one treatment.
-    #[error("samples differ in path, owner or identity")]
-    MismatchedTreatment,
-    /// The two intervals share instants.
-    #[error("intervals {0:?} and {1:?} overlap; nested or concurrent work is not an addend")]
-    OverlappingIntervals(Interval, Interval),
-    /// A summed counter exceeded `u64`.
-    #[error("summed measurement overflows u64")]
-    Overflow,
-    /// The cell is not declared in the ledger.
-    #[error("cell {0:?} is not declared in the ledger")]
-    UnknownCell(String),
-}
 #[cfg(test)]
 pub(crate) fn test_recorder() -> metrics::NoopRecorder {
     metrics::NoopRecorder
 }
+
 #[cfg(test)]
-mod tests {
-    use std::io::{Read, Write};
-    use std::net::{IpAddr, Ipv4Addr, TcpListener, TcpStream};
-    use std::sync::atomic::{AtomicBool, Ordering};
-    use std::thread;
-    use std::time::{Duration, Instant};
-
-    use super::*;
-
-    // MetricsServer::bind installs a process-global recorder. Serialize only
-    // these server tests so another test cannot change the recorder between
-    // the occupied-bind precondition and its assertion. Production is unchanged.
-    static SERVER_TEST_LOCK: Mutex<()> = Mutex::new(());
-
-    #[test]
-    fn process_start_is_recorded_once_and_uptime_advances() {
-        use std::thread;
-        use std::time::Duration;
-
-        record_process_start(Instant::now());
-        let first = process_start();
-        let before = process_uptime();
-
-        thread::sleep(Duration::from_millis(25));
-        // A second record must not move the clock — earliest-wins is what
-        // makes the node, rather than the latest caller, own the origin.
-        record_process_start(Instant::now());
-
-        assert_eq!(
-            process_start(),
-            first,
-            "a second record_process_start must not move the start instant"
-        );
-        let Some(before) = before else {
-            panic!("uptime must be observable after record_process_start");
-        };
-        let after = process_uptime().unwrap_or_else(|| panic!("uptime must stay observable"));
-        assert!(
-            after >= before + Duration::from_millis(20),
-            "uptime must advance with elapsed time: {before:?} -> {after:?}"
-        );
-    }
-
-    #[test]
-    fn warnings_keep_first_message_and_report_in_kind_order() {
-        let warnings = Warnings::new();
-        assert!(warnings.messages().is_empty());
-
-        assert!(warnings.set(WarningKind::ClockOutOfSync, "clock out of sync"));
-        // Core ignores a later message for an already-active kind.
-        assert!(!warnings.set(WarningKind::ClockOutOfSync, "superseded message"));
-        assert!(warnings.set(WarningKind::FatalInternalError, "fatal"));
-        assert!(warnings.set(WarningKind::UnknownNewRulesActivated, "unknown rules"));
-
-        assert_eq!(
-            warnings.messages(),
-            [
-                "unknown rules".to_owned(),
-                "clock out of sync".to_owned(),
-                "fatal".to_owned(),
-            ],
-            "kernel warnings must sort before node warnings regardless of set order"
-        );
-
-        assert!(warnings.unset(WarningKind::ClockOutOfSync));
-        assert!(!warnings.unset(WarningKind::ClockOutOfSync));
-        assert_eq!(
-            warnings.messages(),
-            ["unknown rules".to_owned(), "fatal".to_owned()]
-        );
-    }
-
-    fn unused_ephemeral() -> SocketAddr {
-        SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), 0)
-    }
-
-    /// Identity every test process serves; one recorder, one identity.
-    fn identity() -> EvidenceIdentity {
-        EvidenceIdentity {
-            binary_sha256: Sha256Hex([1; 32]),
-            version: "test".into(),
-            config_sha256: Sha256Hex([2; 32]),
-            corpus: None,
-            backend: "memory".into(),
-            durability: "checkpoint-only".into(),
-            hardware: "test x1".into(),
-        }
-    }
-
-    fn scrape(addr: SocketAddr) -> (u16, String) {
-        let mut last = None;
-        for _ in 0..50 {
-            match TcpStream::connect_timeout(&addr, Duration::from_millis(100)) {
-                Ok(mut stream) => {
-                    stream
-                        .write_all(b"GET /metrics HTTP/1.1\r\nHost: 127.0.0.1\r\nConnection: close\r\n\r\n")
-                        .unwrap_or_else(|error| panic!("write scrape request: {error}"));
-                    stream
-                        .flush()
-                        .unwrap_or_else(|error| panic!("flush scrape request: {error}"));
-                    let mut body = String::new();
-                    stream
-                        .read_to_string(&mut body)
-                        .unwrap_or_else(|error| panic!("read scrape response: {error}"));
-                    let status = body
-                        .split_whitespace()
-                        .nth(1)
-                        .and_then(|token| token.parse().ok())
-                        .unwrap_or(0);
-                    return (status, body);
-                }
-                Err(error) => last = Some(error),
-            }
-            thread::sleep(Duration::from_millis(20));
-        }
-        panic!("scrape connect failed: {last:?}");
-    }
-
-    #[test]
-    fn occupied_address_bind_errors_and_in_process_retry_succeeds() {
-        let _guard = SERVER_TEST_LOCK.lock();
-        let shutdown = Arc::new(AtomicBool::new(false));
-        let occupied = TcpListener::bind(unused_ephemeral())
-            .unwrap_or_else(|error| panic!("occupy port: {error}"));
-        let addr = occupied
-            .local_addr()
-            .unwrap_or_else(|error| panic!("occupied local addr: {error}"));
-        let installed_before = PROMETHEUS_HANDLE.lock().is_some();
-
-        let first = start_metrics(Some(addr), Arc::clone(&shutdown), &identity());
-        assert!(first.is_err(), "occupied bind must fail");
-        assert_eq!(
-            PROMETHEUS_HANDLE.lock().is_some(),
-            installed_before,
-            "failed bind must not install the process recorder"
-        );
-
-        drop(occupied);
-        let server = start_metrics(Some(unused_ephemeral()), shutdown, &identity())
-            .unwrap_or_else(|error| panic!("retry after occupied bind: {error}"))
-            .unwrap_or_else(|| panic!("metrics server"));
-        metrics::counter!("node_metrics_retry_probe").increment(1);
-        let (status, body) = scrape(server.local_addr());
-        assert_eq!(status, 200);
-        assert!(
-            body.contains("node_metrics_retry_probe"),
-            "retry scrape missing recorded metric: {body}"
-        );
-        server.join();
-    }
-
-    #[test]
-    fn scrape_returns_prometheus_text_with_recorded_metrics() {
-        let _guard = SERVER_TEST_LOCK.lock();
-        let shutdown = Arc::new(AtomicBool::new(false));
-        let server = MetricsServer::bind(unused_ephemeral(), shutdown, &identity())
-            .unwrap_or_else(|error| panic!("bind metrics: {error}"));
-        metrics::counter!("node_metrics_scrape_probe").increment(1);
-        let (status, body) = scrape(server.local_addr());
-        assert_eq!(status, 200);
-        assert!(
-            body.contains("text/plain"),
-            "content-type must be prometheus text: {body}"
-        );
-        assert!(
-            body.contains("node_metrics_scrape_probe"),
-            "body must include recorded metric: {body}"
-        );
-        server.join();
-    }
-
-    #[test]
-    fn two_sequential_servers_in_one_process_both_serve() {
-        let _guard = SERVER_TEST_LOCK.lock();
-        let shutdown = Arc::new(AtomicBool::new(false));
-        let first = MetricsServer::bind(unused_ephemeral(), Arc::clone(&shutdown), &identity())
-            .unwrap_or_else(|error| panic!("first: {error}"));
-        metrics::counter!("node_metrics_sequential_probe").increment(1);
-        let (status, body) = scrape(first.local_addr());
-        assert_eq!(status, 200);
-        assert!(body.contains("node_metrics_sequential_probe"));
-        first.join();
-
-        let second = MetricsServer::bind(unused_ephemeral(), shutdown, &identity())
-            .unwrap_or_else(|error| panic!("second: {error}"));
-        metrics::counter!("node_metrics_sequential_probe").increment(1);
-        let (status, body) = scrape(second.local_addr());
-        assert_eq!(status, 200);
-        assert!(body.contains("node_metrics_sequential_probe"));
-        second.join();
-    }
-
-    #[test]
-    fn shutdown_exits_the_listener_thread() {
-        let _guard = SERVER_TEST_LOCK.lock();
-        let shutdown = Arc::new(AtomicBool::new(false));
-        let server = MetricsServer::bind(unused_ephemeral(), Arc::clone(&shutdown), &identity())
-            .unwrap_or_else(|error| panic!("bind metrics: {error}"));
-        let addr = server.local_addr();
-        shutdown.store(true, Ordering::Release);
-        server.join();
-        TcpListener::bind(addr)
-            .unwrap_or_else(|error| panic!("port released after listener join: {error}"));
-    }
-
-    #[test]
-    fn run_retries_metrics_bind_after_occupied_address() {
-        let _guard = SERVER_TEST_LOCK.lock();
-        let shutdown = Arc::new(AtomicBool::new(false));
-        let occupied =
-            TcpListener::bind(unused_ephemeral()).unwrap_or_else(|error| panic!("occupy: {error}"));
-        let busy = occupied
-            .local_addr()
-            .unwrap_or_else(|error| panic!("busy addr: {error}"));
-        assert!(
-            start_metrics(Some(busy), Arc::clone(&shutdown), &identity()).is_err(),
-            "run-path bind must fail on an occupied address"
-        );
-        drop(occupied);
-        let server = start_metrics(Some(unused_ephemeral()), shutdown, &identity())
-            .unwrap_or_else(|error| panic!("run-path retry: {error}"))
-            .unwrap_or_else(|| panic!("server"));
-        let (status, _) = scrape(server.local_addr());
-        assert_eq!(status, 200);
-        server.join();
-    }
-}
+mod tests;
